Changed Footer
Included Social Media links, tried to make color fit the rest of the
page. Still need to stick in the logo but I don't have the file for
that. One problem is that the position of the footer for all the pages
is a little bit off but that has to do with how the partials are
positioned to each other but I can't find the code for that.